+++ src/Resources/contao/dca/tl_page.php
@@ -117,8 +117,12 @@ class  tl_page_om_backend extends Backend
         // import backend user
         $this->import('BackendUser', 'User');
 
+        // get layout
+        $objLayout = \LayoutModel::findByPk($arrPage['layout']);
+        $strTitle = sprintf($GLOBALS['TL_LANG']['tl_page']['layout'][1], '"' . $objLayout->name . '"');
+
         // return link or image only
         return ($this->User->hasAccess('layout', 'themes') && $arrPage['layout'] > 0) ? '<a href="' . $this->addToUrl('do=themes&table=tl_layout&act=edit&id=' . $arrPage['layout'], true,
-                ['do']) . '" title="' . StringUtil::specialchars($title) . '"' . $attributes . '>' . Image::getHtml($icon, $label) . '</a> ' : Image::getHtml(preg_replace('/\.svg$/i', '.svg', $icon)) . ' ';
+                ['do']) . '" title="' . StringUtil::specialchars($strTitle) . '"' . $attributes . '>' . Image::getHtml($icon, $label) . '</a> ' : Image::getHtml(preg_replace('/\.svg$/i', '.svg', $icon)) . ' ';
     }
 }