+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 1.6.7 (2020-01-14)
 
+ + Added: Add name of layout to title on layout button (see #44)
  + Fixed: Version feature breaks version info in sysinfo tab packages (see #46)
  + Fixed: Version info shows lower version than current (see #47) 
  + Fixed: Missing error message if no md-File found (see #43)